vulnerable_version/scada-backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
PyYAML